requirements.txt
Django==3.1
djangorestframework==3.11.1
django-cors-headers==3.4.0
dj-rest-auth==1.1.1
django-allauth==0.42.0
